1.has completed the data parsing about Technology API 2.resolve a error when the amount of money is double ,the app will   go wrong 3.plan to use ViewGroup instead of ViewPager

+++ app/src/main/java/com/huangkun/perfectmemory/activity/AllMoneyActivity.java
@@ -29,7 +29,7 @@ public class AllMoneyActivity extends FragmentActivity {
         ModelDB modelDB = ModelDB.getInstance(this);
         mMoneys = modelDB.loadMoney();
         if (mMoneys.size() != 0){
-            int allNumber = 0;
+            double allNumber = 0;
             for (Money m : mMoneys){
                 allNumber = allNumber + m.getMoneyAmount();
             }
